Formula mrk42(Ardiuno Sketch Ver: A_HardDays_Night
I have commented the code out and entered the formula mk42 (the 23rd actual attempt) Testing with this formula is so far going well. Dry runs indicate successful conversions but have yet to be used with real values. I think if this doesn't work then the calibration is out in the formula, but the bases is right of how its calculated. Thinking of adding an Easter egg as they are awesome. Version A_HardDays_Night